Change policy format for Docker-independent namespaces
Instead of specific/$dockerreference, use transports/$transport/$policyconfigurationscope

This will allow transport-specific policies in the future.

This changes the data structures and parsing code, but mostly not yet the
functionality; everything is looked up through transports/"docker"/... ,
and Docker-reference-less transports are not yet supported.

The only new functionality is that there is a transport-level default
scope (transports/docker/""); at the moment it is redundant with the
global default, but soon it will not be so.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>