Add an expectedSize parameter to PutBlob
Submit this size in docker: destinations, validate it in dir: and oci:.

Preserve the size, if provided by ImageSource, when copying blobs.

Also document that PutBlob must be called before PutManifest and
PutSignatures; this is clearly the necessary ordering if the layer
tarballs are being generated on the fly, or if the manifest destination
wants to reject manifests referring to nonexistent blobs.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>